# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API za praćenje kompetencija",
        "description": "Dokumentacija obuhvata rute za upravljanje kompetencijama, verifikacijama i autentifikaciju korisnika.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Ažuriranje sopstvenog profila",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pera"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Peric"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pera.izmenjen@gmail.com"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Novi opis profila"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "novaLozinka123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil uspešno ažuriran"
                    },
                    "403": {
                        "description": "Možete menjati samo svoj profil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Trajno brisanje korisničkog naloga",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nalog obrisan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/public-profile/{token}": {
            "get": {
                "tags": [
                    "Public"
                ],
                "summary": "Pregled javnog profila (bez logovanja)",
                "operationId": "1dc1b77d4359efda0fec6eacb4af89ab",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešan prikaz javnih podataka"
                    },
                    "404": {
                        "description": "Profil nije pronađen"
                    }
                }
            }
        },
        "/api/users/generate-share-link/{id}": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Generiši UUID link za deljenje profila",
                "operationId": "9581979c586f1777efecff526786cd69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link generisan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Detaljan profil korisnika za Admina/Moderatora",
                "operationId": "e2be8e9382aaf97edefc468ed6d4733e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Svi podaci o korisniku uključujući sve kompetencije"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin ažuriranje bilo kog korisnika",
                "operationId": "97afbcd975ccb1784cdd676dd7398102",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "moderator"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Korisnik uspešno ažuriran od strane admina"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}